created security config,added authorization filter

+++ src/main/java/com/alekhnovich/maxim/testyellowsoftproject/repositories/RunRepository.java
@@ -3,5 +3,5 @@ package com.alekhnovich.maxim.testyellowsoftproject.repositories;
 import com.alekhnovich.maxim.testyellowsoftproject.models.Run;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RunRepository extends JpaRepository<Run,Long> {
+public interface RunRepository extends JpaRepository<Run,Integer> {
 }
