+++ src/main/java/com/alekhnovich/maxim/testyellowsoftproject/services/RunService.java
@@ -2,5 +2,5 @@ package com.alekhnovich.maxim.testyellowsoftproject.services;
 
 import com.alekhnovich.maxim.testyellowsoftproject.models.Run;
 
-public interface RunService extends CrudService<Run,Long> {
+public interface RunService extends CrudService<Run,Integer> {
 }
