+++ src/main/java/com/alekhnovich/maxim/testyellowsoftproject/services/UserService.java
@@ -2,6 +2,6 @@ package com.alekhnovich.maxim.testyellowsoftproject.services;
 
 import com.alekhnovich.maxim.testyellowsoftproject.models.User;
 
-public interface UserService extends CrudService<User,Long> {
+public interface UserService extends CrudService<User,Integer> {
     User getUserByLogin(String login);
 }
